added pcf857x plugin update mode from stable and beta repositories

+++ pcf857x_plugin/pcf857x.py
@@ -3,13 +3,15 @@
 # Python 2/3 compatibility imports
 from __future__ import print_function
 
+__version__ = '0.4.0'
+
 # standard library imports
 import json
 import subprocess
 from sys import modules
 import time
 import platform
-
+from urllib.request import urlopen, Request
 
 blockedPlugin=False
 #import smbus
@@ -18,8 +20,6 @@ try:
 except ModuleNotFoundError:
     blockedPlugin=True
     pass
-
-
 
 # local module imports
 from blinker import signal
@@ -37,7 +37,8 @@ urls.extend(
         u"/pcf857xj", u"plugins.pcf857x.settings_json",
         u"/pcf857xu", u"plugins.pcf857x.update",
         u"/pcf857xt", u"plugins.pcf857x.test",
-        u"/pcf857x_scan", u"plugins.pcf857x.scan"
+        u"/pcf857x_scan", u"plugins.pcf857x.scan",
+        u"/pcf857x_upd", u"plugins.pcf857x.plugupdate"
     ]
 )
 # fmt: on
@@ -53,17 +54,46 @@ if platform.machine() == "armv6l"  or platform.machine() == "armv7l":  # may be 
     demo_mode = False
 
 
+def check_pcf_params(ForceSave):
+    MustSave = False
+    if not ("adr" in pcf):
+        MustSave = True
+        print("adr not found in dict")
+        pcf[u"adr"] = [u""] * gv.sd[u"nbrd"]
+        pcf[u"adr"][0] = u"0x20"
+    if not ("bus" in pcf):
+        MustSave = True
+        print("bus not found in dict")
+        pcf[u"bus"] = "1"
+    if not ("ictype" in pcf):
+        MustSave = True
+        print("ictype not found in dict")
+        pcf[u"ictype"] = "pcf8574"
+    if not ("repo" in pcf):  # repo: S --> Stable or B --> Beta
+        MustSave = True
+        print("repo not found in dict")
+        pcf[u"repo"] = "S"
+    if not ("debug" in pcf):
+        MustSave = True
+        print("debug not found in dict")
+        pcf[u"debug"] = "0"
+    if (MustSave or ForceSave):
+        with open(u"./data/pcf857x.json", u"w") as f:
+            json.dump(pcf, f, indent=4)
+    return
+
 # Read in the pcfadr for this plugin from it's JSON file
 def load_pcfadr():
     global pcf
+    MustSave = False
     try:
         with open(u"./data/pcf857x.json", u"r") as f:
             pcf = json.load(f)  # Read the pcfadr from file
     except IOError:  #  If file does not exist create file with defaults.
-        pcf = {u"adr": [u""] * gv.sd[u"nbrd"], u"bus": 1, u"ictype":"pcf8574",u"debug":"0"}
-        pcf[u"adr"][0] = u"0x20"
-        with open(u"./data/pcf857x.json", u"w") as f:
-            json.dump(pcf, f, indent=4)
+        MustSave = True
+        #pcf = {u"adr": [u""] * gv.sd[u"nbrd"], u"bus": 1, u"ictype":"pcf8574",u"repo":"S",u"debug":"0"}
+        #pcf[u"adr"][0] = u"0x20"
+    check_pcf_params(MustSave)
     return
 
 
@@ -72,6 +102,7 @@ load_pcfadr()
 gv.use_gpio_pins = False
 
 pcf[u"devices"] = {}
+pcf[u"version"] = __version__
 
 if blockedPlugin:
     pcf[u"warnmsg"] = "Unable to load library. please follow instructions on the help page"
@@ -110,7 +141,6 @@ def on_zone_change(name, **kw):
             sid = b * 8 + s  # station index in gv.srvals           
             if gv.output_srvals[sid]:  # station is on
                 byte = byte ^ (1 << s) # use exclusive or to set station bit to 0
-        #print("adding byte: ", hex(byte))
         i2c_bytes.append(byte)
 
     for s in range(gv.sd[u"nbrd"]):
@@ -139,7 +169,7 @@ zones = signal(u"zone_change")
 zones.connect(on_zone_change)
 
 ################################################################################
-# Web pages:                                                                   #
+# Web pages:    needs to be regoranised, too much wrong stuff here             #
 ################################################################################
 
 
@@ -190,6 +220,11 @@ class update(ProtectedPage):
             pcf[u"ictype"] = qdict[u"ictype"]
         else:
             pcf[u"ictype"] = "pcf8574"
+
+        if u"repo" in qdict:
+            pcf[u"repo"] = qdict[u"repo"]
+        else:
+            pcf[u"repo"] = "S"
 
         if u"debug" in qdict: 
             if qdict[u"debug"]=="on":
@@ -242,6 +277,21 @@ def getDevicesOnBus(busNo):
     return devices
 
 
+def proc_files(uri, remote_file,target_dir):
+    #uri = beta_repo + "/-/raw/" + beta_branch + "/" + remote_file
+    response = urlopen(uri)
+    f_data = response.read()
+    try:
+        f_data = f_data.decode('utf-8')
+        print("writing w file")
+        with open(target_dir + "/" + remote_file, "w") as next_file:
+           next_file.write(f_data)
+    except UnicodeDecodeError:
+        print("writing wb file")
+        with open(target_dir + "/" + remote_file, "wb") as next_file:
+           next_file.write(f_data)
+
+
 
 class scan(ProtectedPage):
     """
@@ -258,3 +308,33 @@ class scan(ProtectedPage):
             pcf[u"devices"] = getDevicesOnBus(int(pcf[u"bus"]))
         return template_render.pcf857x(pcf)
 
+class plugupdate(ProtectedPage):
+    """
+    Plugin update
+    """
+
+    def GET(self):
+        # beta repo file example: https://gitlab.com/seventer/pcf857x_plugin_sip/-/raw/dev-04/pcf857x.py
+        beta_repo = "https://gitlab.com/seventer/pcf857x_plugin_sip"
+        beta_branch = "main"
+        stable_repo = "https://raw.github.com/Dan-in-CA/SIP_plugins/"
+        stable_branch = "master"
+        upd_files = {"pcf857x.py": "plugins", "pcf857x.html": "templates"}
+        global pcf
+        uri = ""
+        
+        #print("Update started for", pcf[u"repo"])
+        if (pcf[u"repo"]=="S"):
+            print("updating from stable")
+            uri = stable_repo + "/" + stable_branch + "/pcf857x_plugin/"
+        else:
+            print("Updating from beta")
+            uri = beta_repo + "/-/raw/" + beta_branch + "/"
+            
+        for p in list(upd_files.keys()):
+            f_uri = uri + p
+            print(p,"->",upd_files[p], " - " + f_uri)
+            proc_files(f_uri,p,upd_files[p])
+            
+        #web.seeother(u"/pcf857x")
+        raise web.seeother(u"/restart")
